remove new_text_to_speech, update old one
