+++ apps/api/test/support/client_mocks/elevenlabs.ex
@@ -3,17 +3,19 @@ defmodule Buildel.ClientMocks.Elevenlabs do
   @behaviour ElevenlabsBehaviour
 
   @impl ElevenlabsBehaviour
-  def synthesize(_text, _api_key) do
-    response = %HTTPoison.AsyncResponse{id: make_ref()}
-    send(self(), %HTTPoison.AsyncStatus{id: response.id})
-    send(self(), %HTTPoison.AsyncHeaders{})
+  def speak!(_api_key, state) do
+    {:ok, state.stream_to} |> IO.inspect()
+  end
 
-    send(self(), %HTTPoison.AsyncChunk{
-      id: response.id,
-      chunk: File.read!("test/support/fixtures/real.mp3")
-    })
+  @impl ElevenlabsBehaviour
+  def generate_speech(pid, _text) do
+    IO.inspect("Generating speech")
 
-    send(self(), %HTTPoison.AsyncEnd{id: response.id})
-    response
+    send(
+      pid,
+      {:audio_chunk, File.read!("test/support/fixtures/real.mp3")} |> IO.inspect()
+    )
+
+    :ok
   end
 end
